split archives in two to have different purge policies
we still want to backup logs and maybe other things, but we don't want
to keep those forever the way we do with the rest of the data.

by sending logs in a set of archives with a different name, we enable
different purge policies.

we also enable purging in the script, something that wasn't done before.